refac: lectureTable분리
